api_gateway/requirements.txt
django==5.0.2
djangorestframework==3.14.0
requests==2.31.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7 